support Sequence deserialization only with kotlinx-serialization and UTF-8
